requirements.txt
Authlib==0.15.5
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
pycparser==2.21
python-decouple==3.5
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
